Add serde support. (#10)
* Forks (and brings in) wmidi with serde support.
* bmidi also drops support for SysEx to improve efficiency.
* Adds serde support for main `Bats` struct.

+++ bats-dsp/src/position.rs
@@ -1,8 +1,10 @@
+use serde::{Deserialize, Serialize};
+
 use crate::sample_rate::SampleRate;
 
 /// Position contains the position within the transport. This includes
 /// a beat and sub_beat component.
-#[derive(Copy, Clone, Default, PartialEq, Eq, PartialOrd, Ord)]
+#[derive(Copy, Clone, Default, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
 pub struct Position {
     /// The beat where the top 32bits represent the beat and the bottom 32 bits represents the sub
     /// beat.
@@ -78,8 +80,8 @@ impl std::ops::AddAssign for Position {
 
 impl std::fmt::Debug for Position {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let whole = self.beat() as f64;
-        let fract = self.sub_beat() as f64 / (1u64 << 32) as f64;
+        let whole = self.beat() as f32;
+        let fract = self.sub_beat() as f32 / (1u64 << 32) as f32;
         let beat = whole + fract;
         f.debug_struct("Position").field("beat", &beat).finish()
     }
